hellolaw-ai/requirements.txt
transformers<5.0.0,>=4.6.0
pandas~=2.2.2
scikit-learn==1.3.0
tqdm==4.66.2
datasets==2.13.1
sentence-transformers==2.2.2
evaluate==0.4.0
fastapi~=0.111.0
pydantic~=2.7.1
uvicorn==0.29.0
peft~=0.10.0
numpy~=1.26.4
python-dotenv~=1.0.1
openai~=1.28.0
